@@ -1,5 +1,5 @@
 module github.com/hookdeck/outpost/sdks/outpost-go
 
-go 1.20
+go 1.22
 
 require github.com/ericlagergren/decimal v0.0.0-20221120152707-495c53812d05
