@@ -5,6 +5,7 @@ package hooks
 import (
 	"context"
 	"errors"
+	"github.com/hookdeck/outpost/sdks/outpost-go/internal/config"
 	"net/http"
 )
 
@@ -24,11 +25,13 @@ type HTTPClient interface {
 }
 
 type HookContext struct {
-	BaseURL        string
-	Context        context.Context
-	OperationID    string
-	OAuth2Scopes   []string
-	SecuritySource func(context.Context) (interface{}, error)
+	SDK              any
+	SDKConfiguration config.SDKConfiguration
+	BaseURL          string
+	Context          context.Context
+	OperationID      string
+	OAuth2Scopes     []string
+	SecuritySource   func(context.Context) (interface{}, error)
 }
 
 type BeforeRequestContext struct {
@@ -70,6 +73,11 @@ type Hooks struct {
 	afterSuccessHook  []afterSuccessHook
 	afterErrorHook    []afterErrorHook
 }
+
+var _ sdkInitHook = (*Hooks)(nil)
+var _ beforeRequestHook = (*Hooks)(nil)
+var _ afterSuccessHook = (*Hooks)(nil)
+var _ afterErrorHook = (*Hooks)(nil)
 
 func New() *Hooks {
 	h := &Hooks{
